+++ astropy/io/misc/asdf/tags/transform/basic.py
@@ -28,6 +28,18 @@ class TransformType(AstropyAsdfType):
         if 'bounding_box' in node:
             model.bounding_box = yamlutil.tagged_tree_to_custom_tree(node['bounding_box'], ctx)
 
+        if "inputs" in node:
+            if model.n_inputs == 1:
+                model.inputs = (node["inputs"],)
+            else:
+                model.inputs = tuple(node["inputs"])
+
+        if "outputs" in node:
+            if model.n_outputs == 1:
+                model.outputs = (node["outputs"],)
+            else:
+                model.outputs = tuple(node["outputs"])
+
         return model
 
     @classmethod
@@ -61,6 +73,9 @@ class TransformType(AstropyAsdfType):
             else:
                 bb = [list(item) for item in model.bounding_box]
             node['bounding_box'] = yamlutil.custom_tree_to_tagged_tree(bb, ctx)
+        if type(model.__class__.inputs) != property:
+            node['inputs'] = model.inputs
+            node['outputs'] = model.outputs
 
     @classmethod
     def to_tree_transform(cls, model, ctx):