+++ CHANGES.rst
@@ -79,6 +79,11 @@ astropy.modeling
 - New function ``fix_inputs`` to generate new models from others by fixing
   specific inputs variable values to constants. [#9135]
 
+- ``inputs`` and ``outputs`` are now model instance attributes, and ``n_inputs``
+  and ``n_outputs`` are class attributes. Backwards compatible default values of
+  ``inputs`` and ``outputs`` are generated. ``Model.inputs`` and ``Model.outputs``
+  are now settable which allows renaming them on per user case. [#9298]
+
 
 astropy.nddata
 ^^^^^^^^^^^^^^